CHORE: Added school type, and 1 EKSU Faculty
