+++ data/universities/obafemi-awolowo-university.json
@@ -3,6 +3,7 @@
   "current_vc": "Professor Adebayo Simeon Bamire",
   "name": "Obafemi Awolowo University",
   "acronym": "OAU",
+  "type": "federal",
   "location": {
     "city": "Ile-Ife",
     "state": "Osun"
