+++ data/universities/university-of-lagos.json
@@ -3,6 +3,7 @@
   "current_vc": "",
   "name": "University of Lagos",
   "acronym": "UNILAG",
+  "type": "federal",
   "location": {
     "city": "Lagos",
     "state": "Lagos"
